requirements.txt
faiss-cpu
sentence-transformers
spacy==3.7.2
nltk
pymongo
python-dotenv